[GPIO info] Split ESP8266/ESP32 implementation
